refactor : RandomNumberGenerator 리팩토링 및 테스트 코드 추가
