+++ src/test/java/study/step4/domain/strategy/RandomMoveStrategyTest.java
@@ -1,12 +1,11 @@
 package study.step4.domain.strategy;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import study.step4.util.TestMoveStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("랜덤 이동 전략 테스트")
 class RandomMoveStrategyTest {
